% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paths.R
\name{backend_config_file}
\alias{backend_config_file}
\title{Get backend config file path}
\usage{
backend_config_file()
}
\value{
Path to the config file storing backend executable location.
}
\description{
Get backend config file path
}
