% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paths.R
\name{backend_default_dir}
\alias{backend_default_dir}
\title{Get default backend directory}
\usage{
backend_default_dir()
}
\value{
Path to the default backend installation directory.
}
\description{
Get default backend directory
}
