% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paths.R
\name{backend_default_exe}
\alias{backend_default_exe}
\title{Get default backend executable path}
\usage{
backend_default_exe()
}
\value{
Path to the default converter executable.
}
\description{
Get default backend executable path
}
