% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/backend.R
\name{backend_release_url}
\alias{backend_release_url}
\title{Backend release URL}
\usage{
backend_release_url()
}
\value{
URL to the backend zip file hosted on GitHub Releases.
}
\description{
Backend release URL
}
