% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/backend.R
\name{ensure_html2pdf_backend}
\alias{ensure_html2pdf_backend}
\title{Ensure backend is available}
\usage{
ensure_html2pdf_backend(ask = interactive(), verbose = FALSE)
}
\arguments{
\item{ask}{Logical; ask before downloading when interactive.}

\item{verbose}{Logical; show progress messages.}
}
\value{
Path to the backend executable.
}
\description{
Ensure backend is available
}
\keyword{internal}
