% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paths.R
\name{find_html2pdf_backend}
\alias{find_html2pdf_backend}
\title{Find configured backend executable}
\usage{
find_html2pdf_backend()
}
\value{
Path to backend executable, or an empty string if not found.
}
\description{
Find configured backend executable
}
