% Generated by r6-generator-maven-plugin: do not edit by hand

\name{html_converter}
\alias{html_converter}
\title{Create a new HtmlConverter }
\usage{html_converter(
	fontfiles,
	update
)
}
\arguments{
	\if{html}{\out{<div class="arguments">}}
	\item{fontfiles}{fontfiles - a character vector of font files that will be imported into the converter. - (defaulting to `systemfonts::system_fonts()$path`) - (java expects a RCharacterVector)}
	\item{update}{update - (defaulting to `FALSE`) - (java expects a RLogical)}
	\if{html}{\out{</div>}}
}
\value{
R6 HtmlConverter object: 

}
\description{
for creating PDF and PNG files from HTML. In general this will be created automatically.
but if you have specific fonts you want to use then you may need to pass them to this
function and specify the result in the `converter` parameter of the main functions.
}
\examples{
conv = html2pdfr::html_converter()
}
\keyword{java api}


