% Generated by r6-generator-maven-plugin: do not edit by hand

\name{HtmlConverter}

\alias{HtmlConverter}

\title{HtmlConverter}

\arguments{
	\if{html}{\out{<div class="arguments">}}
	\item{fontfiles}{fontfiles - (java expects a String)}
	\if{html}{\out{</div>}}
}


\description{
missing description

Version: 0.4.2

Generated: 2022-09-24T10:58:50.501
}

\details{
	no details
}

\examples{
## -----------------------------------
## Construct new instance of HtmlConverter
## -----------------------------------
\dontrun{
J = html2pdfr::JavaApi$get()
# appropriate parameter values must be provided
instance = J$HtmlConverter$new(fontfiles)
}
}

\keyword{java api}

\section{Methods}{
	\subsection{Constructors}{
		\itemize{
			\item \href{#method-new}{\code{J$HtmlConverter$new(fontfiles)}}
		}
	}
	\subsection{Class methods}{
		\itemize{
			\item{none}
			\item \code{instance$clone()}
			\item \code{instance$print()}
		}
	}

	\if{html}{\out{<hr>}}
	\if{html}{\out{<a id="method-new"></a>}}
	
	\subsection{Constructor \code{new()}}{
the default no-args constructor

		\subsection{Usage}{
			\if{html}{\out{<div class="r">}}
			\preformatted{
J = html2pdfr::JavaApi$get()
instance = J$HtmlConverter$new(fontfiles);
  			}
			\if{html}{\out{</div>}}
		}
	
		\subsection{Arguments}{
			\if{html}{\out{<div class="arguments">}}
			\itemize{
				\item{fontfiles}{ - (java expects a String)}
			}
			\if{html}{\out{</div>}}
		}

		\subsection{Returns}{
the new R6 HtmlConverter object
		}

	}

}