% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/app.R
\name{launch_html2pdf_app}
\alias{launch_html2pdf_app}
\title{Launch GUI for HTML/XHTML to PDF conversion}
\usage{
launch_html2pdf_app()
}
\value{
Opens a Tk GUI for file selection and conversion.
}
\description{
Launch GUI for HTML/XHTML to PDF conversion
}
