% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paths.R
\name{set_html2pdf_backend}
\alias{set_html2pdf_backend}
\title{Set backend executable path}
\usage{
set_html2pdf_backend(path)
}
\arguments{
\item{path}{Path to converter_backend.exe.}
}
\value{
Invisibly returns the normalized backend path.
}
\description{
Set backend executable path
}
